Moved internal #defines to C file.  Fixed #defines to use unsigned int consts.
For header file:

Commented insert/remove as they are not yet verified.  Please use insert_nc
and remove_nc for now (probably a better idea in any event for the applications
we are verifying at this point).

Added capacity function.

Removed free_pq, but it's reasonable to add it back at some point.

Once we have a precise bound from body_sink, we can move on to make().  Should
probably take a capacity rather than the INITIAL_SIZE constant.